+++ cornerstone-common/src/main/java/com/cornerstone/common/core/ErrorCode.java
@@ -13,6 +13,10 @@ public enum ErrorCode implements IErrorCode {
     FORBIDDEN(403, "无权限访问"),
     /** 资源不存在 */
     NOT_FOUND(404, "资源不存在"),
+    /** 请求方法不支持 */
+    METHOD_NOT_ALLOWED(405, "请求方法不支持"),
+    /** 不支持的媒体类型 */
+    UNSUPPORTED_MEDIA_TYPE(415, "不支持的媒体类型"),
     /** 系统内部错误 */
     INTERNAL_ERROR(500, "系统内部错误");